CA-235 Add product to compare product page

+++ app/src/main/java/cenergy/central/com/pwb_store/realm/RealmController.java
@@ -136,7 +136,9 @@ public class RealmController {
             @Override
             public void execute(@NonNull Realm realm) {
                 RealmResults<CompareProduct> realmCompareProducts = realm.where(CompareProduct.class).equalTo(CompareProduct.FIELD_SKU, sku).findAll();
-                realmCompareProducts.deleteAllFromRealm();
+                if (realmCompareProducts != null) {
+                    realmCompareProducts.deleteAllFromRealm();
+                }
             }
         });
